+++ d6/game_magic_weapon.h
@@ -8,9 +8,9 @@ struct MagicWeapon : Sprite {
 	constexpr static float cSpeedMin{ 50.f / gDesign.fps };
 	constexpr static float cSpeedMax{ 200.f / gDesign.fps };
 	constexpr static float cSpeedInc{ 10.f / gDesign.fps };
-	constexpr static float cFlyRadianMin{ M_PI * 1 / gDesign.fps };
-	constexpr static float cFlyRadianMax{ M_PI * 3 / gDesign.fps };
-	constexpr static float cAroundRadiansInc{ M_PI * 2 / gDesign.fps };
+	constexpr static float cFlyRadianMin{ float(M_PI * 1 / gDesign.fps) };
+	constexpr static float cFlyRadianMax{ float(M_PI * 3 / gDesign.fps) };
+	constexpr static float cAroundRadiansInc{ float(M_PI * 2 / gDesign.fps) };
 
 	float speed{ cSpeedMin }, flyRadian{};
 	bool following{};
@@ -24,7 +24,7 @@ struct MagicWeapon : Sprite {
 		owner = owner_;
 		pos = bornPos;
 		frames = &gLooper.frames_magicWeapon;
-		frameIndex = index;
+		frameIndex = (float)index;
 		body.SetRotate(radians).SetAnchor(cAnchor);
 	}
 
@@ -91,7 +91,7 @@ struct MagicWeapon : Sprite {
 				rb = -rb;
 				following = true;
 				while (true) {									// begin rotate fly around the owner
-					for (float r = rb, re = rb + M_PI * 2; r < re; r += cAroundRadiansInc) {
+					for (float r = rb, re = float(rb + M_PI * 2); r < re; r += cAroundRadiansInc) {
 						if (target) goto LabBegin;
 						if (!owner) goto LabEnd;
 						radians = -r;
